// src/app/shared/model/lehengaMeasure.model.ts
import { MeasurementStyle } from './measurementStyle.model';
import { MeasurementValue } from './measurementValue.model';

export class LehengaMeasure {
    aroundWaist: [MeasurementValue];
    aroundHip: [MeasurementValue];
    lehengaLength: [MeasurementValue];
    lehengaClosingSide: [string];
    lehengaClosingWith: [string];
}